feat: [sc-167] Link Generator Update

+++ src/app/pages/applications/applications-list/modules/grant-application/grant-application.component.ts
@@ -5,6 +5,7 @@ import { ApplicationRead } from 'src/app/interfaces/application.interface';
 import { TransformApplicationForList, compare, transformApplicationForList } from 'src/app/utilities/application.utils';
 import { ToastrService } from 'ngx-toastr';
 import { HttpErrorResponse } from '@angular/common/http';
+import * as _ from '../../../../../utilities/globals';
 
 type TableObj = {
     list: TransformApplicationForList;
@@ -65,7 +66,7 @@ export class GrantApplicationComponent implements OnInit {
         NgbdSortableHeaderDirective<TransformApplicationForList>
     >;
 
-    constructor(private applicationService: ApplicationService, private toastr: ToastrService) {}
+    constructor(private applicationService: ApplicationService, private toastr: ToastrService) { }
 
     ngOnInit() {
         this.handleFetchApplication();
@@ -116,40 +117,53 @@ export class GrantApplicationComponent implements OnInit {
     }
 
     remove(application: TransformApplicationForList[0]) {
-        this.applicationService.destroy(application.applicationPk).subscribe({
-            next: (data: any) => {
-                const status = data?.status;
-                if (status) {
-                    this.grantApplication.urgentGrants.list = this.grantApplication.urgentGrants.list.filter(
-                        (item) => item.applicationPk !== application.applicationPk
-                    );
-                    this.grantApplication.submissions.list = this.grantApplication.submissions.list.filter(
-                        (item) => item.applicationPk !== application.applicationPk
-                    );
-                    this.grantApplication.grantsTeamReview.list = this.grantApplication.grantsTeamReview.list.filter(
-                        (item) => item.applicationPk !== application.applicationPk
-                    );
-                    this.grantApplication.advisersReview.list = this.grantApplication.advisersReview.list.filter(
-                        (item) => item.applicationPk !== application.applicationPk
-                    );
-                    this.grantApplication.dueDiligence.list = this.grantApplication.dueDiligence.list.filter(
-                        (item) => item.applicationPk !== application.applicationPk
-                    );
-                    this.grantApplication.budgetReviewAndFinalization.list =
-                        this.grantApplication.budgetReviewAndFinalization.list.filter(
+        _.confirmMessage({
+            title: '<strong>Are you sure you want to delete this application?</strong>',
+            icon: 'question',
+            buttons: {
+                showClose: true,
+                showCancel: true,
+                focusConfirm: false
+            },
+            confirmButtonText: '<i class="fa fa-trash"></i> Delete',
+            cancelButtonText: '<i class="fa fa-thumbs-down"></i> No, cancel'
+        }, () => {
+            this.applicationService.destroy(application.applicationPk).subscribe({
+                next: (data: any) => {
+                    const status = data?.status;
+                    if (status) {
+                        this.grantApplication.urgentGrants.list = this.grantApplication.urgentGrants.list.filter(
                             (item) => item.applicationPk !== application.applicationPk
                         );
-                    this.grantApplication.financialManagementCapacity.list =
-                        this.grantApplication.financialManagementCapacity.list.filter(
+                        this.grantApplication.submissions.list = this.grantApplication.submissions.list.filter(
                             (item) => item.applicationPk !== application.applicationPk
                         );
-                }
-            },
-            error: (err: HttpErrorResponse) => {
-                const errorMessage = err?.error?.message ? `message: ${err?.error?.message}` : '';
-                const statusCode = err?.status ? `status: ${err?.status}` : '';
-                this.toastr.error(`Error trying to remove application. ${statusCode} ${errorMessage} `);
-            },
+                        this.grantApplication.grantsTeamReview.list = this.grantApplication.grantsTeamReview.list.filter(
+                            (item) => item.applicationPk !== application.applicationPk
+                        );
+                        this.grantApplication.advisersReview.list = this.grantApplication.advisersReview.list.filter(
+                            (item) => item.applicationPk !== application.applicationPk
+                        );
+                        this.grantApplication.dueDiligence.list = this.grantApplication.dueDiligence.list.filter(
+                            (item) => item.applicationPk !== application.applicationPk
+                        );
+                        this.grantApplication.budgetReviewAndFinalization.list =
+                            this.grantApplication.budgetReviewAndFinalization.list.filter(
+                                (item) => item.applicationPk !== application.applicationPk
+                            );
+                        this.grantApplication.financialManagementCapacity.list =
+                            this.grantApplication.financialManagementCapacity.list.filter(
+                                (item) => item.applicationPk !== application.applicationPk
+                            );
+                    }
+                },
+                error: (err: HttpErrorResponse) => {
+                    const errorMessage = err?.error?.message ? `message: ${err?.error?.message}` : '';
+                    const statusCode = err?.status ? `status: ${err?.status}` : '';
+                    this.toastr.error(`Error trying to remove application. ${statusCode} ${errorMessage} `);
+                },
+            });
         });
+
     }
 }
